test: move elasticsearch setup fixture to parent conftest
